Moved sphinx dependencies from inmanta-dev-dependencies to requirements.dev.txt and dropped unused extensions (PR #6330)
# Description

The `sphinx` extra for `inmanta-dev-dependencies` is currently only used by `inmanta-core` and the `developerdocs`. Given that the indirect dependencies occasionally gives conflicts when dependabot tries to bump versions, it's easier to depend on them directly and drop them from `inmanta-dev-dependencies`.

I additionally checked which of the sphinx dependencies I think we actually use and I dropped `recommonmark`, `sphinx-tab` and `sphinxcontrib-serializinghtml` because I don't think we use them anymore.

PRs for `developerdocs` and `inmanta-dev-dependencies` will follow.