Allow product's conf.py to inject version information when building docs (PR #6339)
This allows the product to inject the version information before this file is even loaded. This in turn allows us to use this version for other purposes. This was not possible before because e.g. for the iso product the version would be inaccurate (`INMANTA_DONT_DISCOVER_VERSION` -> `1.0.0`) until after this file had finished loading. By allowing version injection, the version becomes accurate at load-time.

See inmanta/inmanta-service-orchestrator#410