Fix test_override_inmanta_package test. (PR #6523)
# Description

This test checks that an exception is raised when trying to override the inmanta-core package in an env.
The regex that was used always assumed that the order of the conflicting packages would be:

`inmanta-core==4.0.0` and then `inmanta-core=.*`

The recent major version bump on core from 9.x.x to 10.x.x changed the reported error message (because of lexicographic order I assume ?) to be:

`inmanta-core=.*` and then `inmanta-core==4.0.0`

which was the root cause of the failure

This PR makes sure we check for the error with both orderings.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)