Fixed reference to OpenAPI docs to work from any page (PR #6654)
# Description

Fixes reference to OpenAPI docs so that it works when clicked on from nested pages. I added a sphinx file for the page so that we can refer to it using sphinx' native linking. The generated html will eventually be overrided by the redoc generated OpenAPI docs. I ran the docs build with this change and the resulting `openapi.html` was identical.

closes #6338