Port inmanta to pydantic v2 (PR #6739)
# Description

Migrate to Pydantic v2. This is the final PR, which is a follow-up from the PoC PRs #6463 and #6685.

Current status: I still need to refine a few details (marked with TODO), and review the full mypy diff (it's big). But I think any changes from here on out should be relatively minor so it makes sense to open it for review now. inmanta-lsm PR will follow soon.

part of #6462

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)