update dependabot (PR #6858)
# Description

Set up dependabot for the new branches:

make sure dependabot is enabled for master and disabled for the iso branch that moved to patch because
it should only receive security updates (which we manage by hand for now, https://github.com/inmanta/irt/issues/512).
Merge the config change (if any) using the merge tool, which acts as a guard against accidental branch divergence.

on the product repo and repos that were diverged earlier in the major release process (inmanta-lsm), additionally
enable updates for the newly created iso branch but set it up to ignore all extensions.

- [ ] Changelog entry