Fixed lexing of matching keyword (PR #7022)
# Description

Turns out the new implementation of the lexer for the `matching` keyword was a bit too eager. Indeed, I now see that the [PLY lexer docs](https://ply.readthedocs.io/en/latest/ply.html#lex) state "You should avoid writing individual rules for reserved words. [If you do] those rules will be triggered for identifiers that include those words as a prefix such as “forget” or “printed”. This is probably not what you want."

This is breaking some tests on the solutions team so if approved in the morning before I start working, feel free to just trigger the merge tool.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)