Add tracing support (PR #7615)
# Description

This adds the groundwork for tracing to core. It does not add any tracing to the test cases (yet).

Documentation and a convention on how to tag are also still required.

There is a startup overhead for all the instrumentation. This resulted in test_app.py to be slower. To not make it the run longer and to make the tests more readable, it reworked test_app.py The term and kill time have been increased to make the tests run stable on jenkins

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)