event propagation bugfix in scheduler (PR #8514)
# Description

Bugfix in event propagation. We had a special case for sending events where we set `deploying=set()` in order to force the receiving resource to be scheduled, even if it happens to already be running. But this also hid any other resources that are running. In other words, if another dependency of the receiving resource happens to be running, it becomes invisible and the receiving resource will not wait for it. It will then see that resource in the available state and skip.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)