Changed default for InmantaBootloader's logging config argument (PR #8530)
# Description

This new parameter was breaking all extension pipelines because they call the bootloader without arguments. The current implementation then breaks later on when `get_current_instance()` is called.

I'm not sure if this is the proper solution, but it's the cheapest to get the builds unstuck until it can be properly considered after the holidays.

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)