added restore state setting for agent restart (Issue #8601, PR #8610)
# Description

Added environment setting to toggle whether the scheduler should restore its state from the database (or alternatively fall back to increment calculation). This setting is introduced exclusively as an easy fallback mechanism in case of support calls related to scheduler state inconsistencies, so we don't keep building on them. We are confident in the scheduler correctness, but we'd rather be prepared just in case.

closes #8601

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [ ] Changelog entry
- [ ] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [ ] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)