Fix issue where the e2e tests fail to install the std module. (PR #8826)
# Description

Fix issue where the e2e tests fail to install the std module. These tests didn't need python code from the std module before, but since std version `8.3.0` they do because that version added a reference definition. When a reference is defined, the python code is always exported (for now) independent from whether the reference is used or not.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~
- [ ] ~~If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)~~