Added support to provide the access token of the `GET /api/v2/docs` endpoint via the `token` query parameter instead of via the Authorization header. (PR #9046)
# Description

**Same PR as https://github.com/inmanta/inmanta-core/pull/9030 but applied on the iso8 branch due to a merge conflict**

Added support to provide the access token of the `GET /api/v2/docs` endpoint via the `token` query parameter instead of via the Authorization header. This change is required to give the web-console the possibility to set the authorization token. Currently, the Authorization header is not set because this is a link from the web-console application to an external application (swagger). A similar PR will be required on the lsm extension for its documentation endpoint.

Part of https://github.com/inmanta/inmanta-lsm/issues/2096

# Self Check:

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [x] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] ~~If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)~~