Fix bug where the JWT configuration is not cleared after reloading the configuration files. (PR #9094)
# Description

The `AuthJWTConfig` class keeps a cache with the JWT configuration. This configuration is derived on the configuration in the Config class. If the Config class is updated, the AuthJWTConfig class should be reset so that it picks up the new configuration. This reset didn't happen, which mainly caused issues for the test suite. This PR fixes that problems.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~
- [ ] ~~If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)~~