Change endpoints that specifically fetch the status of a resource, to only use the rps table (PR #9031)
# Description

Part1 of this PR was already merged but was reverted due to an error on the inmanta-lsm test suite. [I fixed the error in this PR.](https://github.com/inmanta/inmanta-lsm/pull/2110)

In this first pass, I wanted to address the queries that explicitly fetched the status from the resource table. Meaning that the status we return was directly tied to the query. These were the affected queries:

1. Get_status_for (Removed, not being used)
2. Get_resource_ids_with_status (Removed. Only used on mark_deployed.)
3. Get_undeployable (Removed, it was not being used)
4. Get_resource_states_latest_version (used on the trigger_get_status endpoint to check if the state of the resources in the DB corresponds to the internal state of the scheduler and returns the discrepancies
5. Get_current_resource_state (part of the stable api, used to return the state of a resource in the latest version of the configuration model or None if it is not present)
6. Get_resource_deploy_summary (used on resource_list to output the deploy summary)
7. ResourceView (used on resource_list)
8. get_metric_value (used for the environment metrics)

On a next pass I'll address the queries where the status is incidentally returned, i.e. we are returning SELECT * FROM resource . In these instances I'll first try to assert how it is being used, and if we are using that method to check the status of a resource or are we more interested in its other fields

[Passed the full test suite](https://jenkins.inmanta.com/job/core/job/inmanta-core/job/change-more-queries-to-fecth-status-from-rps-table/15/)

closes #9080 9080

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [ ] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [ ] Code is clear and sufficiently documented
- [ ] No (preventable) type errors (check using make mypy or make mypy-diff)
- [ ] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )
- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)